Move comparator override specific functions into comparator module
